Refactoring & Fix highlighting
