+++ src/main/kotlin/com/github/izhangzhihao/rainbow/brackets/RainbowHighlighter.kt
@@ -6,16 +6,21 @@ import com.intellij.codeInsight.daemon.impl.HighlightInfoType
 import com.intellij.lang.annotation.HighlightSeverity
 import com.intellij.openapi.editor.DefaultLanguageHighlighterColors
 import com.intellij.openapi.editor.colors.EditorColorsManager
+import com.intellij.openapi.editor.colors.EditorColorsScheme
 import com.intellij.openapi.editor.colors.TextAttributesKey
 import com.intellij.openapi.editor.colors.TextAttributesScheme
 import com.intellij.openapi.editor.markup.EffectType
 import com.intellij.openapi.editor.markup.TextAttributes
 import com.intellij.psi.PsiElement
+import com.intellij.ui.JBColor
+import com.intellij.util.ui.UIUtil
 import org.jetbrains.annotations.TestOnly
 import java.awt.Color
 import java.awt.Font
 
 object RainbowHighlighter {
+
+    val DEFAULT_KOTLIN_LABEL_COLOR = JBColor(0x4a86e8, 0x467cda)
 
     const val NAME_ROUND_BRACKETS = "Round Brackets"
     const val NAME_SQUARE_BRACKETS = "Square Brackets"
@@ -135,4 +140,36 @@ object RainbowHighlighter {
                         .range(element)
                         .create()
             }
+
+
+    private val KEY_HTML_CODE = TextAttributesKey.createTextAttributesKey("HTML_CODE")
+    private val KEY_KOTLIN_LABEL = TextAttributesKey.createTextAttributesKey("KOTLIN_LABEL")
+    private val KEY_MATCHED_BRACE_ATTRIBUTES =
+            TextAttributesKey.createTextAttributesKey("MATCHED_BRACE_ATTRIBUTES")
+
+
+    fun fixHighlighting() {
+        val globalScheme = EditorColorsManager.getInstance().globalScheme
+        globalScheme.setInherited(KEY_HTML_CODE, !settings.isRainbowifyHTMLInsideJS)
+
+        // kotlin label
+        val kotlinLabelColor = DEFAULT_KOTLIN_LABEL_COLOR.takeUnless { settings.isRainbowifyKotlinLabel }
+        val kotlinLabel = TextAttributes(kotlinLabelColor, null, null, EffectType.BOXED, Font.PLAIN)
+        globalScheme.setAttributes(KEY_KOTLIN_LABEL, kotlinLabel)
+
+        // matched brace
+        val matchedBraceAttributes = if (settings.isOverrideMatchedBraceAttributes) {
+            TextAttributes(null, JBColor(0x99ccbb, 0x3b514d), null, EffectType.BOXED, Font.BOLD)
+        } else {
+            val isDark = UIUtil.isUnderDarcula()
+            val foregroundColor = if (isDark) Color(0xffef28) else null
+            val fontType = if (isDark) Font.BOLD else Font.PLAIN
+            TextAttributes(foregroundColor, JBColor(0x99ccff, 0x3b514d), null, EffectType.BOXED, fontType)
+        }
+        globalScheme.setAttributes(KEY_MATCHED_BRACE_ATTRIBUTES, matchedBraceAttributes)
+    }
+
+    private fun EditorColorsScheme.setInherited(key: TextAttributesKey, inherited: Boolean) {
+        setAttributes(key, if (inherited) TextAttributes.USE_INHERITED_MARKER else TextAttributes())
+    }
 }
