+++ src/main/kotlin/com/github/izhangzhihao/rainbow/brackets/settings/RainbowSettings.kt
@@ -21,7 +21,6 @@ class RainbowSettings : PersistentStateComponent<RainbowSettings> {
     var isDoNOTRainbowifyBracketsWithoutContent = false
     var version = "Unknown"
     var isRainbowifyHTMLInsideJS = true
-    var isRainbowifyKotlinFunctionLiteralBracesAndArrow = true
     var isRainbowifyKotlinLabel = true
     var isOverrideMatchedBraceAttributes = true
